% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chesmore2011-data.R
\docType{data}
\name{chesmore2011}
\alias{chesmore2011}
\title{Coding Matrix from Chesmore (2011)}
\format{Matrix}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(chesmore2011)
}
\description{
Coding matrix used by....
}
\examples{
data(chesmore2011)
t <- tdsc(wave, c_matrix=chesmore2011)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{CodingMatrix}
\keyword{TDSC}
