% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyBands.R
\name{emptyBands}
\alias{emptyBands}
\title{Empty Bands Discovery}
\usage{
emptyBands(...)
}
\arguments{
\item{...}{Two or more TDSC objects}
}
\description{
Identifies unused codes across multiple S-matricies
}
\examples{
emptyBands()

}
\keyword{TDSC}
