% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farr2007-data.R
\docType{data}
\name{farr2007}
\alias{farr2007}
\title{Coding Matrix from Farr (2007)}
\format{Matrix}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(farr2007)
}
\description{
Coding matrix used by Farr in his 2007 PhD thesis.
}
\examples{
data(farr2007)
t <- tdsc(wave, c_matrix=farr2007)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{CodingMatrix}
\keyword{TDSC}
