% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.a.matrix.R
\name{normalise.a.matrix}
\alias{normalise.a.matrix}
\title{Normalise an A Matrix}
\usage{
normalise.a.matrix(td, method = "scale")
}
\arguments{
\item{td}{A tdsc object}

\item{method}{Either scale (default) or codewords}
}
\description{
Normalises the A Matrix of a tdsc object either by scale or by codewords.
}
\examples{
library(tuneR)
wave <- readWave(system.file("extdata", "1.wav", package="tdsc"))
t <- tdsc(wave)
t <- normalise.a.matrix(t)
t <- normalise.a.matrix(t, method="codewords")

}
