% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdsc-class.R
\docType{class}
\name{tdsc-class}
\alias{tdsc-class}
\title{An S4 class to hold results from TDSC}
\description{
An S4 class to hold results from TDSC
}
\section{Slots}{

\describe{
\item{\code{raw}}{Two column vector of Durations and Shapes}

\item{\code{positive}}{Identify non-negative sections}

\item{\code{codelist}}{Vector of sequential epoch codings}

\item{\code{b_matrix}}{The basic matrix}

\item{\code{c_matrix}}{The coding matrix}

\item{\code{s_matrix}}{The S-matrix}

\item{\code{a_matrix}}{The A-matrix}

\item{\code{sample_count}}{The number of samples in the waveform}

\item{\code{epoch_count}}{The number of identified epochs}

\item{\code{stdsc}}{Statistical TDSC feature vector}
}}

