% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tdsc.plot}
\alias{tdsc.plot}
\title{Plot the A matrix}
\usage{
tdsc.plot(td, plotter = "persp")
}
\arguments{
\item{td}{A TDSC object}

\item{plotter}{Function used to plot the A matrix (persp or perp3D)}
}
\description{
Plot the A matrix from a tdsc analysis.
}
\examples{
\dontrun{
tdsc.plot(td)
tdsc.plot(td, plotter="persp3D")
}

}
