% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdsc.w.R
\name{tdsc.w}
\alias{tdsc.w}
\title{TDSC analysis comaptible with windowing}
\usage{
tdsc.w(start, wave = NULL, window.length = NULL)

tdsc.w(start, wave = NULL, window.length = NULL)
}
\arguments{
\item{start}{Start position in samples}

\item{wave}{Wave object to analyse}

\item{window.length}{Length of window}
}
\description{
Wrapper function for TDSC analysis that is compatable with the windowing function
of the package sonicscrewdriver.
}
\examples{
\dontrun{
windowing(wave, window.length=1000, window.overlap=0, FUN=tdsc.w)
}
}
