% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tdsc_plot}
\alias{tdsc_plot}
\title{Plot the A matrix or S Matrix}
\usage{
tdsc_plot(td, plotter = "persp", ...)
}
\arguments{
\item{td}{A TDSC object}

\item{plotter}{Function used to plot the A matrix (persp or perp3D)  or S matrix (hist)}

\item{...}{Parameters to pass to plotting function}
}
\description{
Plot the A matrix or S Matrix from a tdsc analysis.
}
\examples{
\dontrun{
tdsc.plot(td)
tdsc.plot(td, plotter="persp3D")
tdsc.plot(td, plotter="hist")
}

}
