% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dmatrix.R
\name{c2dmatrix}
\alias{c2dmatrix}
\title{Convert a coding matrix to a distributed matrix}
\usage{
c2dmatrix(input, sf = 100)
}
\arguments{
\item{input}{A tdsc object or a matrix}

\item{sf}{The scaling factor}
}
\description{
Identifies unused codes across multiple S-matricies
}
\examples{
c2dmatrix()

}
\keyword{TDSC}
