% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2dmatrix.R
\name{c2dmatrix}
\alias{c2dmatrix}
\title{Convert a coding matrix to a distributed matrix}
\usage{
c2dmatrix(t, sf = 100)
}
\arguments{
\item{t}{A tdsc object or a matrix}

\item{sf}{The scaling factor}
}
\description{
This function converts a coding matrix of any kind into a distributed matrix as described in Farr (2007).
}
\examples{
c2dmatrix(as.matrix(c(1,2,3,4), nrow=2))

}
\references{
Farr (2007) “Automated Bioacoustic Identification of Statutory Quarantined Insect Pests”. PhD thesis. University of York.
}
\keyword{TDSC}
