% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyBands.R
\name{emptyBands}
\alias{emptyBands}
\title{Empty Bands Discovery}
\usage{
emptyBands(...)
}
\arguments{
\item{...}{Two or more TDSC objects}
}
\description{
Identifies unused codes across multiple S-matricies. Unused bands can be used to reduce the codebook as in Stammers (2011).
}
\examples{
library(tuneR)
wave <- readWave(system.file("extdata", "1.wav", package="tdsc"))
t <- tdsc(wave)
emptyBands(t,t)

}
\references{
Stammers (2011) “Audio Event Classification for Urban Soundscape Analysis”. PhD thesis. University of York.
}
\keyword{TDSC}
