% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farr2007-data.R
\docType{data}
\name{farr2007}
\alias{farr2007}
\title{Coding Matrix from Farr (2007)}
\format{Matrix}
\usage{
data(farr2007)
}
\description{
Coding matrix used by Farr in his 2007 PhD thesis.
}
\examples{
data(farr2007)
t <- tdsc(wave, c_matrix=farr2007)
}
\references{
Farr (2007) “Automated Bioacoustic Identification of Statutory Quarantined Insect Pests”. PhD thesis. University of York.
}
\keyword{CodingMatrix}
\keyword{TDSC}
